IBP-7142 use kotlin-csv instead of manual parser
